requirements.txt
fastapi==0.115.5
uvicorn[standard]
sqlalchemy[asyncio]==2.0.*
asyncpg
alembic
anthropic
openpyxl==3.1.5
fpdf2>=2.8.0
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
slowapi
structlog
pydantic>=2.0
pydantic-settings
pypdf>=4.0
